woot

+++ world_counter.py
@@ -5,7 +5,7 @@ def ncr(n, r):
     r = min(r, n-r)
     numer = reduce(op.mul, range(n, n-r, -1), 1)
     denom = reduce(op.mul, range(1, r+1), 1)
-    return numer / denom
+    return int(numer / denom)
 
 f = open("chart.txt")
 world_count = 0;
